new test, separated integration tests
